backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]==2.5.3
supabase==2.7.4
python-dotenv==1.0.0
websockets==13.1